There are models with too many columns and translated and they are slow prefetching them

and they are even called in website checkout process where it should be faster than possible

and the prefetched values are not even used

e.g. Reading 'product_tmpl_id' column the query called is

    SELECT "product_template"."id" as "id",
        COALESCE("product_template__website_meta_title"."value", "product_template"."website_meta_title") as "website_meta_title",
        COALESCE("product_template__website_meta_description"."value", "product_template"."website_meta_description") as "website_meta_description",
        COALESCE("product_template__website_meta_keywords"."value", "product_template"."website_meta_keywords") as "website_meta_keywords",
        COALESCE("product_template__seo_name"."value", "product_template"."seo_name") as "seo_name",
        COALESCE("product_template__name"."value", "product_template"."name") as "name",
        COALESCE("product_template__description"."value", "product_template"."description") as "description",
        COALESCE("product_template__description_purchase"."value", "product_template"."description_purchase") as "description_purchase",
        COALESCE("product_template__description_sale"."value", "product_template"."description_sale") as "description_sale",
        COALESCE("product_template__button_text"."value", "product_template"."button_text") as "button_text",
        COALESCE("product_template__quantity_title"."value", "product_template"."quantity_title") as "quantity_title",
        ...
    FROM "product_template"
    LEFT JOIN
        "ir_translation" AS "product_template__website_meta_title"
        ON ("product_template"."id" = "product_template__website_meta_title"."res_id"
        AND "product_template__website_meta_title"."type" = 'model'
        AND "product_template__website_meta_title"."name" = 'product.template,website_meta_title'
        AND "product_template__website_meta_title"."lang" = 'en_US'
        AND "product_template__website_meta_title"."value" != '')
    LEFT JOIN
        "ir_translation" AS "product_template__website_meta_description"
        ON ("product_template"."id" = "product_template__website_meta_description"."res_id"
        AND "product_template__website_meta_description"."type" = 'model'
        AND "product_template__website_meta_description"."name" = 'product.template,website_meta_description'
        AND "product_template__website_meta_description"."lang" = 'en_US'
        AND "product_template__website_meta_description"."value" != '')
    LEFT JOIN
        "ir_translation" AS "product_template__website_meta_keywords"
        ON ("product_template"."id" = "product_template__website_meta_keywords"."res_id"
        AND "product_template__website_meta_keywords"."type" = 'model'
        AND "product_template__website_meta_keywords"."name" = 'product.template,website_meta_keywords'
        AND "product_template__website_meta_keywords"."lang" = 'en_US'
        AND "product_template__website_meta_keywords"."value" != '')
    LEFT JOIN
        "ir_translation" AS "product_template__seo_name"
        ON ("product_template"."id" = "product_template__seo_name"."res_id"
        AND "product_template__seo_name"."type" = 'model'
        AND "product_template__seo_name"."name" = 'product.template,seo_name'
        AND "product_template__seo_name"."lang" = 'en_US'
        AND "product_template__seo_name"."value" != '')
    LEFT JOIN
        "ir_translation" AS "product_template__name"
        ON ("product_template"."id" = "product_template__name"."res_id"
        AND "product_template__name"."type" = 'model'
        AND "product_template__name"."name" = 'product.template,name'
        AND "product_template__name"."lang" = 'en_US'
        AND "product_template__name"."value" != '')
    LEFT JOIN
        "ir_translation" AS "product_template__description"
        ON ("product_template"."id" = "product_template__description"."res_id"
        AND "product_template__description"."type" = 'model'
        AND "product_template__description"."name" = 'product.template,description'
        AND "product_template__description"."lang" = 'en_US'
        AND "product_template__description"."value" != '')
    LEFT JOIN
        "ir_translation" AS "product_template__description_purchase"
        ON ("product_template"."id" = "product_template__description_purchase"."res_id"
        AND "product_template__description_purchase"."type" = 'model'
        AND "product_template__description_purchase"."name" = 'product.template,description_purchase'
        AND "product_template__description_purchase"."lang" = 'en_US'
        AND "product_template__description_purchase"."value" != '')
    LEFT JOIN
        "ir_translation" AS "product_template__description_sale"
        ON ("product_template"."id" = "product_template__description_sale"."res_id"
        AND "product_template__description_sale"."type" = 'model'
        AND "product_template__description_sale"."name" = 'product.template,description_sale'
        AND "product_template__description_sale"."lang" = 'en_US'
        AND "product_template__description_sale"."value" != '')
    LEFT JOIN
        "ir_translation" AS "product_template__button_text"
        ON ("product_template"."id" = "product_template__button_text"."res_id"
        AND "product_template__button_text"."type" = 'model'
        AND "product_template__button_text"."name" = 'product.template,button_text'
        AND "product_template__button_text"."lang" = 'en_US'
        AND "product_template__button_text"."value" != '')
    LEFT JOIN
        "ir_translation" AS "product_template__quantity_title"
        ON ("product_template"."id" = "product_template__quantity_title"."res_id"
        AND "product_template__quantity_title"."type" = 'model'
        AND "product_template__quantity_title"."name" = 'product.template,quantity_title'
        AND "product_template__quantity_title"."lang" = 'en_US'
        AND "product_template__quantity_title"."value" != '')
    WHERE "product_template".id IN (...)

Planning Time: 2.564 ms
Execution Time: 0.334 ms

Using with_context(prefetch_fields=False) the query is

    SELECT "product_template"."id" as "id",
        COALESCE("product_template__name"."value", "product_template"."name") as "name"
    FROM "product_template"
    LEFT JOIN
        "ir_translation" AS "product_template__name"
        ON ("product_template"."id" = "product_template__name"."res_id"
        AND "product_template__name"."type" = 'model'
        AND "product_template__name"."name" = 'product.template,name'
        AND "product_template__name"."lang" = 'en_US'
        AND "product_template__name"."value" != '')
    WHERE "product_template".id IN (...)

Planning Time: 0.285 ms
Execution Time: 0.029 ms

It is 9x faster