@odony 2022-06-28
Have you been able to confirm these kinds of benefits with more significant volumes and timings? 9x improvement on sub-millisecond timings isn't very meaningful, it's easily drowned in the noise.

Also it's a bit risky to manually disable prefetching for those fields:
1) no guarantee that a module won't break your fragile optimization at any time
2) the trade-off may be negative if you're causing extra queries, may not be a net win in some cases

BTW ci/runbot is red, which may be related to this (2).

We might want to consider turning off prefetching for translatable fields, but I'm really not sure this is gonna be a good idea, because it's not always a positive trade-off. On average, my gut feeling is that it's better to do an extra JOIN rather than risk needing extra queries. 

Besides, this problem might disappear entirely as of v16 when translations get inlined in table rows, and when the shop pages get cached more aggressively. So it may not be worth a lot of trouble at this point.

/cc @rco-odoo